tests refactored to /tests dir
