+++ frontend/jest.config.js
@@ -25,7 +25,6 @@ module.exports = {
   testPathIgnorePatterns: [
     '/.next/',
     '/node_modules/',
-    '/tests/',
     '/coverage/',
   ],
   transform: {
