+++ frontend/package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "axios": "^0.21.1",
+    "enzyme": "^3.11.0",
     "http-proxy-middleware": "^1.0.6",
     "next": "10.0.5",
     "react": "17.0.1",
@@ -28,7 +29,6 @@
     "babel-jest": "^26.6.3",
     "babel-plugin-styled-components": "^1.12.0",
     "cross-env": "^7.0.3",
-    "enzyme": "^3.11.0",
     "eslint": "^7.18.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-plugin-import": "^2.22.1",
@@ -37,6 +37,5 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "express": "^4.17.1",
     "jest": "^26.6.3"
-  },
-  "proxy": "http://localhost:5000"
+  }
 }
